renaming and code cleanup
