+++ src/main/java/com/kpabr/backrooms/util/BiomeListBuilder.java
@@ -7,22 +7,14 @@ import java.util.*;
 import java.util.stream.Stream;
 
 public class BiomeListBuilder {
-    private final TreeMap<Double, RegistryKey<Biome>> biomeList = new TreeMap<>();
+    private final HashMap<LevelParameters, RegistryKey<Biome>> biomeList = new HashMap<>();
 
-    /**
-     *
-     * @param chance takes double value from 0.0 to 1.0, if less than 0.0 or more than 1.0 it will convert value to 2.0
-     */
-    public BiomeListBuilder addBiome(RegistryKey<Biome> biome, double chance) {
-        if(chance > 1 || chance < 0) {
-            this.biomeList.put(2.0, biome);
-        } else {
-            this.biomeList.put(chance, biome);
-        }
+    public BiomeListBuilder addBiome(RegistryKey<Biome> biome, LevelParameters parameters) {
+        this.biomeList.put(parameters, biome);
         return this;
     }
 
-    public TreeMap<Double, RegistryKey<Biome>> getBiomeList() {
+    public HashMap<LevelParameters, RegistryKey<Biome>> getBiomeList() {
         return this.biomeList;
     }
