+++ src/main/java/com/kpabr/backrooms/util/BiomeRegistryList.java
@@ -4,19 +4,43 @@ import net.minecraft.util.registry.Registry;
 import net.minecraft.util.registry.RegistryEntry;
 import net.minecraft.world.biome.Biome;
 
-import java.util.TreeMap;
+import java.util.HashMap;
 import java.util.stream.Stream;
 
 public class BiomeRegistryList {
     public static final Double DEFAULT_CHANCE_VALUE = 2.D; // Use for default biome type
-    private final TreeMap<Double, RegistryEntry<Biome>> biomeList = new TreeMap<>();
+    private final HashMap<LevelParameters, RegistryEntry<Biome>> biomeList = new HashMap<>();
 
     public Stream<RegistryEntry<Biome>> getBiomeEntries() {
         return this.biomeList.values().stream();
     }
 
-    public RegistryEntry<Biome> findNearest(double key) {
-        return biomeList.ceilingEntry(key).getValue();
+    public RegistryEntry<Biome> findNearest(LevelParameters params) {
+        
+        LevelParameters mostSimilarKey = null;
+        double smallestDistance = Float.MAX_VALUE;
+
+        for (LevelParameters key : biomeList.keySet()) {
+            double distance = calculateDistance(key, params);
+            distance = distance * key.rareness;
+            if (distance < smallestDistance) {
+                smallestDistance = distance;
+                mostSimilarKey = key;
+            }
+        }
+
+        return biomeList.get(mostSimilarKey);
+    }
+
+    private double calculateDistance(LevelParameters key, LevelParameters params) {
+        // Calculate Euclidean distance between key values and params from the input
+        double sum = 0;
+        sum += Math.pow(key.temperature - params.temperature, 2);
+        sum += Math.pow(key.moistness - params.moistness, 2);
+        sum += Math.pow(key.integrity - params.integrity, 2);
+        sum += Math.pow(key.purity - params.purity, 2);
+        sum += Math.pow(key.toxicity - params.toxicity, 2);
+        return (double) Math.sqrt(sum);
     }
 
     public static BiomeRegistryList from(Registry<Biome> biomeRegistry, BiomeListBuilder biomeList) {
